Created Sta class, to split number crunching from argument handling and printing. Will help with unit testing
